Fix DNS
When trying to connect the DB instance through the SQL developer and the corporate machine DNS might be configured with their own domain. And suppose if we want to connect the client DB instances then we may face problem to connect thorugh SQL developer or through any DB client.Solution is is fix the DNS values.